+++ CHANGELOG.md
@@ -21,10 +21,12 @@ Wherever possible, a reference to an issue in the project issue tracker should b
   [#93](https://gitlab.data.bas.ac.uk/MAGIC/assets-tracking-service/-/issues/93)
 * Release/version script and deployments from main branch
   [#96](https://gitlab.data.bas.ac.uk/MAGIC/assets-tracking-service/-/issues/96)
-* New DB migration script to automatically create/number files
+* Script to create DB migrations
   [#107](https://gitlab.data.bas.ac.uk/MAGIC/assets-tracking-service/-/issues/107)
 * `data run` command combining `data fetch` and `data export` commands, replacing cron script logic
   [#105](https://gitlab.data.bas.ac.uk/MAGIC/assets-tracking-service/-/issues/105)
+* Script to clean up prerelease venvs
+  [#115](https://gitlab.data.bas.ac.uk/MAGIC/assets-tracking-service/-/issues/115)
 
 ### Fixed